Update accumulated usage using batched calls and then group the calls
Improve the performance of accumulator by batching the accumulated usage
updates function call and then by grouping usage submissions in a batch
using a key formed with resource instance and time period.

In a real world scenario, we are not expecting multiple usage submissions
for a resource instance and a time period at any given point in time.
So in this case, this would not improve the performance of the accumulator.
However, if we do get multiple usage submissions for any reason, this would
improve the performance of accumulator.