Update rated usage using batched calls and then group the calls
Improve the performance of rating by batching the rated usage update function
calls and then by grouping usage submissions in a batch using a key formed
with organization and time period.

On a sample local perf test of 40 orgs, 40 instances and 40 usage submissions,
with a total of 64,000, Reduced the end to end processing time by more than
2/3rd.

Related to [#101021756] at Pivotal Tracker.