Fix individual response headers of a batch request
Using /batch endpoint through a batch request would return empty headers.

HTTP responses from middlewares with header.Location are converted to an array
of response objects with header.Location and the array is sent back to a client
as a response body of the batch request.

At the client side, this array of response objects needs to be translated back
to individual responses with correct location header for each response.

Fixes [Finishes #102122388] at Pivotal Tracker.