Set maximum listeners to unlimited for all EventEmitter instances
EventEmitters print a warning if more than 10 listeners are added for
a particular event. This is a useful default which helps finding memory leaks.

Setting this value to unlimitted to avoid getting this warning when handling
more than 10 requests at a time. In the case of a batch request, we will not
see this warning till we handle more than 9 requests in a batch.

Fixes github issue #82.